update: file structure of auto-drive api
