+++ packages/auto-drive/src/api/calls/read.ts
@@ -1,19 +1,19 @@
 import { ArgsWithoutPagination, ArgsWithPagination } from '../../utils/types'
-import { AutoDriveApi } from '../connection'
 import { PaginatedResult } from '../models/common'
 import { ObjectInformation, ObjectSearchResult, ObjectSummary, Scope } from '../models/objects'
 import { UserInfo } from '../models/user'
+import { AutoDriveApiHandler } from '../type'
 
 /**
  * Retrieves the root objects based on the specified scope.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithPagination<{ scope: Scope }>} query - The query parameters including scope, limit, and offset.
  * @returns {Promise<ObjectSummary[]>} - A promise that resolves to an array of ObjectSummary representing the root objects.
  * @throws {Error} - Throws an error if the request fails.
  */
 export const getRoots = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithPagination<{ scope: Scope }>,
 ): Promise<PaginatedResult<ObjectSummary>> => {
   const response = await api.sendRequest(
@@ -36,13 +36,13 @@ export const getRoots = async (
  * This method sends a request to the server to fetch a list of objects
  * that are shared with the user, based on the specified pagination parameters.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithPagination} query - The query parameters including limit and offset for pagination.
  * @returns {Promise<ObjectSummary[]>} - A promise that resolves to an array of ObjectSummary representing the shared objects.
  * @throws {Error} - Throws an error if the request fails.
  */
 export const getSharedWithMe = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithPagination,
 ): Promise<PaginatedResult<ObjectSummary>> => {
   const response = await api.sendRequest(
@@ -60,7 +60,7 @@ export const getSharedWithMe = async (
 }
 
 export const searchByNameOrCID = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ value: string; scope: Scope }>,
 ): Promise<ObjectSearchResult[]> => {
   const response = await api.sendRequest(
@@ -83,13 +83,13 @@ export const searchByNameOrCID = async (
  * This method sends a request to the server to fetch a list of objects
  * that have been deleted, based on the specified pagination parameters.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithPagination} query - The query parameters including limit and offset for pagination.
  * @returns {Promise<ObjectSummary[]>} - A promise that resolves to an array of ObjectSummary representing the deleted objects.
  * @throws {Error} - Throws an error if the request fails.
  */
 export const getDeleted = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithPagination,
 ): Promise<PaginatedResult<ObjectSummary>> => {
   const response = await api.sendRequest(
@@ -112,13 +112,13 @@ export const getDeleted = async (
  * This method sends a request to the server to fetch details about the
  * object, including its metadata and other relevant information.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string }>} query - The query parameters containing the CID of the object to retrieve.
  * @returns {Promise<ObjectInformation>} - A promise that resolves to the information of the requested object.
  * @throws {Error} - Throws an error if the request fails.
  */
 export const getObject = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string }>,
 ): Promise<ObjectInformation> => {
   const response = await api.sendRequest(`/objects/${query.cid}`, {
@@ -137,13 +137,13 @@ export const getObject = async (
  *
  * This method sends a request to the server to fetch the summary of the object.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string }>} query - The query parameters containing the CID of the object whose upload status is to be retrieved.
  * @returns {Promise<ObjectSummary>} - A promise that resolves to the summary of the requested object.
  * @throws {Error} - Throws an error if the request fails.
  */
 export const getObjectSummary = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string }>,
 ): Promise<ObjectSummary> => {
   const response = await api.sendRequest(`/objects/${query.cid}/summary`, {
@@ -164,13 +164,13 @@ export const getObjectSummary = async (
  * of the object, which can indicate whether the upload is pending, completed,
  * or failed.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string }>} query - The query parameters containing the CID of the object whose upload status is to be retrieved.
  * @returns {Promise<ObjectInformation['uploadStatus']>} - A promise that resolves to the upload status of the requested object.
  * @throws {Error} - Throws an error if the request fails.
  */
 export const getObjectUploadStatus = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string }>,
 ): Promise<ObjectInformation['uploadStatus']> => {
   const response = await api.sendRequest(`/objects/${query.cid}/status`, {
@@ -191,13 +191,13 @@ export const getObjectUploadStatus = async (
  * associated with the object. The owners can provide insights into who
  * has access to or control over the object.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string }>} query - The query parameters containing the CID of the object whose owners are to be retrieved.
  * @returns {Promise<ObjectInformation['owners']>} - A promise that resolves to the list of owners of the requested object.
  * @throws {Error} - Throws an error if the request fails.
  */
 export const getObjectOwners = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string }>,
 ): Promise<ObjectInformation['owners']> => {
   const response = await api.sendRequest(`/objects/${query.cid}/owners`, {
@@ -218,13 +218,13 @@ export const getObjectOwners = async (
  * with the object. The metadata can include various details about the object,
  * such as its name, type, size, and other relevant information.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string }>} query - The query parameters containing the CID of the object whose metadata is to be retrieved.
  * @returns {Promise<ObjectInformation['metadata']>} - A promise that resolves to the metadata of the requested object.
  * @throws {Error} - Throws an error if the request fails.
  */
 export const getObjectMetadata = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string }>,
 ): Promise<ObjectInformation['metadata']> => {
   const response = await api.sendRequest(`/objects/${query.cid}/metadata`, {
@@ -241,11 +241,11 @@ export const getObjectMetadata = async (
 /**
  * Get upload and download limits of the user
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @returns {Promise<UserInfo>} - A promise that resolves to the user info.
  * @throws {Error} - Throws an error if the request fails.
  */
-export const getMe = async (api: AutoDriveApi): Promise<UserInfo> => {
+export const getMe = async (api: AutoDriveApiHandler): Promise<UserInfo> => {
   const response = await api.sendRequest('@me', {
     method: 'GET',
   })
