+++ packages/auto-drive/src/api/calls/write.ts
@@ -1,5 +1,5 @@
-import { ArgsWithoutPagination } from '../../utils/types';
-import { AutoDriveApi } from '../connection';
+import { ArgsWithoutPagination } from '../../utils/types'
+import { AutoDriveApiHandler } from '../type'
 
 /**
  * Shares an object with a specified public ID.
@@ -8,13 +8,13 @@ import { AutoDriveApi } from '../connection';
  * by its CID. The object will be shared with the provided public ID, allowing
  * access to the object for users with that ID.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string; publicId: string }>} query - The query parameters containing the CID of the object to share and the public ID to share it with.
  * @returns {Promise<any>} - A promise that resolves to the response from the server.
  * @throws {Error} - Throws an error if the sharing process fails.
  */
 export const shareObject = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string; publicId: string }>,
 ) => {
   const response = await api.sendRequest(
@@ -42,13 +42,13 @@ export const shareObject = async (
  * by its CID as deleted. This action is typically irreversible and should
  * be used with caution.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string }>} query - The query parameters containing the CID of the object to mark as deleted.
  * @returns {Promise<void>} - A promise that resolves when the object has been marked as deleted.
  * @throws {Error} - Throws an error if the marking process fails.
  */
 export const markObjectAsDeleted = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string }>,
 ): Promise<void> => {
   const response = await api.sendRequest(`/objects/${query.cid}/delete`, {
@@ -67,13 +67,13 @@ export const markObjectAsDeleted = async (
  * by its CID. The restoration process may depend on the server's implementation
  * and the object's current state.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string }>} query - The query parameters containing the CID of the object to restore.
  * @returns {Promise<void>} - A promise that resolves when the object has been successfully restored.
  * @throws {Error} - Throws an error if the restoration process fails.
  */
 export const restoreObject = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string }>,
 ): Promise<void> => {
   const response = await api.sendRequest(`/objects/${query.cid}/restore`, {
@@ -94,13 +94,13 @@ export const restoreObject = async (
  * by its CID. The publication process may depend on the server's implementation
  * and the object's current state.
  *
- * @param {AutoDriveApi} api - The API instance used to send requests.
+ * @param {AutoDriveApiHandler} api - The API instance used to send requests.
  * @param {ArgsWithoutPagination<{ cid: string }>} query - The query parameters containing the CID of the object to publish.
  * @returns {Promise<void>} - A promise that resolves when the object has been successfully published.
  * @throws {Error} - Throws an error if the publication process fails.
  */
 export const publishObject = async (
-  api: AutoDriveApi,
+  api: AutoDriveApiHandler,
   query: ArgsWithoutPagination<{ cid: string }>,
 ): Promise<{ result: string }> => {
   const response = await api.sendRequest(`/objects/${query.cid}/publish`, {
