+++ packages/auto-drive/src/api/index.ts
@@ -1,4 +1,5 @@
 export * from './calls/index'
 export * from './connection'
+export * from './handler'
 export * from './models/index'
 export * from './wrappers'
